Relatively major changes made to allow banded alignment.  Small bugs could still exist (this may always be true though...).
With the --nosmall --banded option combination to cmalign, a
memory-efficient banded, non D&C alignment is performed, which is up
to 2X faster than the D&C --banded alignment (at least for SSU).  

Another new cmalign option, the --full option, forces the alignment
outputted from cmalign to contain all consensus columns of the model,
which is useful if more than one alignment to the same model are to be
merged together.


Finally, the new --bandexpand option provides a naive, inefficient, hacky
solution to a situation that causes banded cmalign to break.  Namely, if the
the length of the sequence being aligned is outside the band range for
the root note (defined by dmin[0] and dmax[0]).  The --bandexpand
option makes cmalign change the bands by expanding them, either by
decreasing dmin[v] by the difference (dmin[0] - length(target)) for
all states v if length(target) < dmin[0], or by increasing dmax[v] by
the difference (length(target) - dmax[0]) if length(target) >
dmax[0].  This is a very experimental approach.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@1428 db3e70e4-12e1-0310-b23f-81a6e3d09ba7