IMPORTANT CHANGE: cmio.c modified to always remove ambiguities involving insert states 1 state prior to an END_E. These states are 'detached' from the rest of the model by setting all transitions into them as 0.0. This is done by a function call withing read_ascii_cm() and read_binary_cm(), and no option exists currently to NOT modify this behavior.
This means that it's currently impossible to identically recreate
output from previous versions of cmalign <cm> (or cmsearch, or
cmscore), if <cm> was built with a cmbuild that does not remove these
ambiguities (cmbuild 0.7 and earlier does not remove these
ambiguities).

Also since the last commit, I reorganized hmmband.c related code a
bit.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@1663 db3e70e4-12e1-0310-b23f-81a6e3d09ba7