Fixed bug i11 involving detaching inserts in zero consensus length hairpins.
I was a little dissappointed to discover that a CP9 cannot be
built that exactly mirrors a CM for these cases (although a CP9 gets
really close to the CM). cp9-test in the testsuite is the only current
executable that actually checks CP9s, and it prints a warning in such
a case. 

The bandcyk_*_test changes are b/c I had previously changed a function
prototype used in those programs. cmsearch.c change is cosmetic, and
the rmark_MPI_cmsearch.pl change is a small fix irrelevant to bug i11.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@1817 db3e70e4-12e1-0310-b23f-81a6e3d09ba7