This commit is not in the main development infernal branch, but rather in the branch that will become release 0.72.
Fixed bug i11 involving detaching inserts in zero consensus length
hairpins.

I was a little dissappointed to discover that a CP9 cannot be
built that exactly mirrors a CM for these cases (although a CP9 gets
really close to the CM). cp9-test in the testsuite is the only current
executable that actually checks CP9s, and it prints a warning in such
a case.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/0.7stable@1818 db3e70e4-12e1-0310-b23f-81a6e3d09ba7