Added new cmsearch behavior for E-values (-E) with the enforce option (--enfseq <s> --enfstart <x>):
o modelconfig.c::ConfigCMEnforce() enforces <s> (this used to occur in
  modelconfig.c::ConfigCM())
o E-values are calculated prior to reparameterizing the model to
  enforce <s>
o search is done with the enforced CM (this hasn't changed)
o cm->enf_scdiff is calc'ed as the *approximate* score difference b/t
  a subparse of <s> through the enforced states of the model before/after
  <s> is enforced by ConfigCMEnforce()
o for each hit, E-values are determined after subtracting
  cm->enf_scdiff bits from the bit score found using the enforced CM.

To achieve this behavior I had to add a new function
modelconfig::ConfigGlobal(), which takes a CM in local mode and
reverts to global (details on motivation in code). This required I add
cm->root_trans a vector of floats that are the transitions for the
ROOT_S state (cm->t[0]). These are saved in
modelconfig.c::ConfigLocal() which zeroes them, and then reset in 
modelconfig.c::ConfigGlobal().


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@1897 db3e70e4-12e1-0310-b23f-81a6e3d09ba7