cmbuild --rsearch is now minimally functional. cmsearch -R (for running RSEARCH searches) is known to be buggy, but I'm not worried about it b/c it will soon be deprecated, hopefully.
cmbuild --rsearch reads the default RIBOSUM matrix (RIBOSUM85-60) in a
different format than the RSEARCH format in that the background (g)
probs are included in that file. The target freqs are back calculated
to get emission probs for the single sequence CM, and these are
logoddsified using the g vector from the RIBOSUM as the null
model. This procedure does not give log odds scores that are identical
to the RIBOSUM matrix, but that's okay. RIBOSUM/RSEARCH are emitted
two sequences, thus have joint emission probs emitting the query and
target at once, while Infernal has conditional emission probs emitting
the target dependent on the query. 

About to benchmark.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@1911 db3e70e4-12e1-0310-b23f-81a6e3d09ba7