User Guide and man pages now updated to my (EPN) satisfaction for release 0.8. Also small code changes that I deemed appropriate while update User Guide were made.
The single trcyk.c change is to make the PrintFancyAli() call
consistent with the new prototype of that function and should not
affect functionality of trcyk at all.

This commit does not include most of the recent edits to the User
Guide (those were in previous commits the past several days).


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@1955 db3e70e4-12e1-0310-b23f-81a6e3d09ba7