This revision reintroduces the cmsearch --fil-Smax-hmm bug that was fixed in revision 2981 (in the 1.0 branch, revision 2982 in the trunk). The reason I'm doing this is to make the 1.0.1 release correspond as closely as possible to the code I used to get the results in my thesis, so they could be reproduced if desired. My thesis (now "published") refers to the 1.0.1 version even though it did not actually exist when the various experiments were performed.
This bug is serious: when cmsearch is run with a calibrated model, and
the appropriate filter threshold from the CM file gives a predicted
survival fraction of less than 1.0 but greater than <x> from
--fil-Smax-hmm <x> (by default, 0.5), then the buggy cmsearch sets the
filter threshold that yields a predicted surv fraction of <x>, when
what it should do is turn the HMM filter off. 

This bug affects exactly 2 families on the rmark-2 benchmark that is
used extensively in my thesis. It actually *improves* performance
(albeit slightly, by 1 MER), but also accelerates the benchmark with
default settings by about 20%. That is, when the bug is fixed the MER
increases by 1 and the benchmark takes ~25 hours instead of ~21 hours
using default cmsearch.

This revision (r2989) should become version 1.0.1. I have run the
benchmark with this code to make sure it matches the results in my
thesis. 

After this commit, I will fix the bug and do a new commit. That fixed
code will eventually become the 1.0.2 release. 

I will not merge this revision (or the next one) to the trunk. 

The only change in this revision is the specific code changes
necessary to reintroduce the bug.

For more details see
~nawrockie/notebook/9_1019_inf_1p01_and_1p02_releases/00LOG.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.0@2989 db3e70e4-12e1-0310-b23f-81a6e3d09ba7