cm_mx.c, hybridsearch.c, cp9_dp.c, funcs.h:
These 4 files were revised to fix a 'bug' involving how null3
penalties are calc'ed for HMM Forward hits. Previously the length of
an HMM hit was always presumed to be W for null3 calculation -- this
resulted in inflated penalties for models where W/clen ratio was large
(like group I introns, RF00028 for which this bug was first detected
by Rfam curator Jen Daub).

The fix is to assume HMM Forward hits are always the consensus length
of the model (not W). I could have used the average length of emitted
sequence from a model but that would have required a lot of code
changes and I judged it was not worth it. See 
~nawrockie/notebook/9_0924_inf_rfam_question/00LOG and
~nawrockie/notebook/9_1019_inf_1p01_and_1p02_releases/00LOG 
for more information on how this bug was found and its fix was
evaluated. 

HMM Forward hit lengths are unknown because a scanning Forward
implementation reports only high scoring end points, which sum over
different possible starting points. In the current implementation
Backward is run beginning at each Forward end point to get a
reasonable start, but only after the null3 penalty is applied and the
resulting bit score is evaluated to see if it exceeds the HMM filter
bit score cutoff, if not, Backward is not run for the hit and it 
does not survive to the next round.

cmcalibrate.c: 
Default maximally useful HMM filter survival fraction
changed from 0.95 to 0.75.

cm_parsetree.c:
Changed verbosity level of a single ESL_DPRINTF statement from a '1'
to a '3'.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.0@2999 db3e70e4-12e1-0310-b23f-81a6e3d09ba7