Bugfix (MSCYK)
Numerical behavior of Bifurcation state:
At BIF-type states, we need to add two values,
both of which are partial alignments and already
include a bias term.  0 represents -infty.
In general, A + -infty cannot be relied upon
to calculate to -infty (fails if A > bias).

One approach would be to convert one of the two
scores to a cost and subtract it, as is done 
for transition and emission scores.  However,
this would limit the effective range of the score
to be flipped to 2 x bias, and could severely
underestimate the score of high-scoring hits.

The solution in this patch does not fix the
underlying problem; rather, it fixes the one
known case where the behavior of A + -infty
was being relied upon.  An array of mask vectors
is applied after calculating the bifurcation,
to re-zero combinations of (j,d) which are
not possible.

This solution is sufficient only if scores of 
-infty are not expected to occur at valid
matrix cells.  Otherwise, 0 will need to be
checked and caught to produce a mask for each
summation of BIFL and BIFR terms, which is
expected to be more computationally expensive
than the current implementation.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3039 db3e70e4-12e1-0310-b23f-81a6e3d09ba7