Bugfixes (MSCYK):
1.  Handling of A + -infty in BIF: now checks
left and right terms for 0 (-infty), produces
a mask, and re-zeroes.  This reverses the design
decision of the previous patch (r2039).

2.  Update wrap-around convenience vectors (d = -1, d = -2)
after calculating BIF.

3.  Sign of initialization on S -> e transition.

4.  Change in meta-model parameters.  Known issue:
if S -> e transition cost is larger in magnitude than
bias, all alignments will be impossible.  This arises
because bias and scaling are determined on the basis
of the CM, and don't account for the meta-model parameters.
Also, if the cost is only slightly smaller than the
bias term, it imposes a noticeable limit on the length
of non-homologous residue segments.

Remaining known issues relate to either parameterization
(as in 4 above), and whether/how to report scores directly
from model states - the requirement to cycle through
S -> SM before recording hits imposes a significant
penalty on scores.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3040 db3e70e4-12e1-0310-b23f-81a6e3d09ba7