cmsearch.c:
Added the --cm-idx <n> and --cm-name <s> options, which allow the user
to specify that only CM number <n> or CM named <s> be used for
searching. The motivation for this was for ssu-align development, to
remove the need for 2 separate CM files for users who want to search
with all 3 default SSU models and those who just want to use one.

An analogous change was previously made to cmalign with revision 3023.

cmsearch.man and cmsearch.sqc: 
Updated to include the --cm-idx and --cm-name options.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.0@3045 db3e70e4-12e1-0310-b23f-81a6e3d09ba7