Major correction in numerical range of byte scores: offset alignment scores by base_b rather than bias_b.
Because base_b is generally much larger than bias_b,
this compounded a range problem in the calculation of
scores at bifurcations.  Both segments are alignments
and therefore already contain the offset term; so we
must subtract one of the offsets from the sum.  If
we sum first, overflow is very likely, and if not
corrected for limits scores to (0, 255-base_b).  If
we subtract base_b first from one of the components,
underflow is likely.

The implementation here catches the overflow events
in a mask, uses unsaturated addition to calculate the
size of the overflow, and (after subtracting the
offset and the transition cost) re-adds this remainder
with saturated math.  See notebook pgs 58-59 for
details.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3054 db3e70e4-12e1-0310-b23f-81a6e3d09ba7