cmalign --matchonly (which outputs an aln with only the clen consensus columns) now only allocates an alignment of size clen*nseq, instead of first creating an alignment with inserts and then stripping them.
Also, the way posterior probabilities are added to an MSA by cmalign
was changed. This is now done in Parsetrees2Alignment() which is
passed ptrs to the 2 posterior code strings, allowing posterior
annotation to work with the --matchonly option.

The cmalign --onepost option was removed, which modified the aln
output to only include a single digit of posterior annotation. (This
single digit posterior annotation was not the same as HMMER3
annotation). cmalign.sqc and cmalign.man were updated to reflect
removal of --onepost.

cmalign --withali <f> now propagates Infernal posterior annotation in
<f> to the final alignment if -p is also enabled. 

cmbuild.c, cmemit.c and cp9_modelmaker.c: very minor changes, only
necessary to make Parsetrees2Alignment() calls match the new function
prototype (postcode1 and postcode2 are passed as NULL).  


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.0@3069 db3e70e4-12e1-0310-b23f-81a6e3d09ba7