Merged changes from 1.0 branch with command:
svn merge -r 3076:3077
https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.0

Log from r3077 in 1.0 branch:
~~~~~~~~~~~~~~~~~~~~
cm_parsetree.c: Parsetrees2Alignment() can now optionally allocate and
fills each aligned sequence (and GR posterior annotation) in the
nascent MSA independently, only as each parsetree is converted to an
aligned sequence/posterior codes. After creation of an aseq and
posterior codes for a given parsetree, the parsetree is freed. This
change reduces the required amount of memory of the function to be no
more than the size of all the parsetrees and ESL_SQs, instead of that
size plus the size of the MSA (which, btw, is about the same size as
the parsetrees if the MSA has about 6*clen positions for a model of
consensus length clen).

cmalign.c's and cmemit.c's Parsetrees2Alignment() call use this more
memory efficient option, but all other calls do not as of yet. Changes
to the other files were nec to be consistent with new
Parsetrees2Alignment() function prototype.
~~~~~~~~~~~~~~~~~~~~


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3078 db3e70e4-12e1-0310-b23f-81a6e3d09ba7