Added possibly temporary fix to problem of a top-level 'make' compiling Easel twice, once for infernal progs and once for H3 progs (see log r3097). A top-level 'make' now moves into 'hmmer/' and executes 'make hmmer_lib' to only build the libhmmer.a library. Building libhmmer.a seems to not require any compiled easel code, so hmmer/easel/ remains unused.
Log r3097's description of how infernal was building was wrong.
Creating a symlink from infernal/hmmer/easel to infernal/easel is not
actually possible and r3097 was in fact building two separate
libeasel.a's, one in each easel dir. 

Mysteriously (to me), the ordering of '-lhmmer -leasel' in MYLIBS
definition of infernal/src/Makefile.in is still necessary (infernal
progs will not build if '-leasel -lhmmer' order is used.  


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3099 db3e70e4-12e1-0310-b23f-81a6e3d09ba7