Brought infernal testsuite up to date with sqc.
Makefile.in and testsuite/*sqc:
Updated usage of sqc so it is consistent with recent changes 
as of r503 in easel trunk.

dev-exercises.sqc:
Removed this file because it is no longer used during a 'make
devcheck'. Instead, a 'make devcheck' calls sqc with each of the
cm*.sqc input files. 

cmcalibrate.sqc:
In an old revision (r3006) I incorrectly put cmcalibrate.sqc in the
top-level directory instead of testsuite/ when manually merging from
the 1.0 branch
(https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.0)


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3163 db3e70e4-12e1-0310-b23f-81a6e3d09ba7