cmalign.c: Made the --p7 option only visible with a --devhelp.
my_p7_msvfilter.c:
Fixed one line in this code which was written about 18 months ago (!)
for experimental H3-based P7 banding, here's the svn diff:

-      } else if (esl_FCompare(P7XMX(i,p7G_J), P7XMX(i-1,p7G_E) + tec, tol) == eslOK) /* note: P7XMX(i-1,p7G_E) differs from Viterbi traceback, where it's P7XMX(i,p7G_E), not sure why */
+      } else if (esl_FCompare(P7XMX(i,p7G_J), P7XMX(i,p7G_E) + tec, tol) == eslOK) 

This was preventing --p7 from working at all in cmalign. This change
is largely irrelevant because --p7 is an experimental option that
isn't even listed with a -h (but it is listed with a --devhelp). For
SSU-ALIGN it is completely irrelevant because that package never uses
P7 banding. The plan is to use P7 banding in future releases of
SSU-ALIGN and possibly cmalign. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3286 db3e70e4-12e1-0310-b23f-81a6e3d09ba7