Two important changes: 1. Changed the default prior to the 'p33' prior, which improves    performance on the rmark3 benchmark over the old default prior. The    old default prior from versions 0.56 through 1.0.2 can be used with    the --p56 option to 'cmbuild'.     Xref: ~nawrockie/notebook/10_0722_inf_prior_reestimation/           ~nawrockie/notebook/10_0830_inf_rmark3_again/           ~nawrockie/notebook/10_0913_lm/
2. By default, the prior probability of the null3 model is now 1/65536
   (1/2^16) instead of 1/32. This lessens the severity of the null3
   correction, and achieves optimal performance over choices of prior
   probabilities for null3 on rmark3 given the new prior, a feature of
   the new prior which makes it 'better' than the old one, in my
   opinion. 
   Xref: ~nawrockie/notebook/10_0722_inf_prior_reestimation/

To achieve (2), a <null3_omega> value has been added to the CM data
structure, which is the prior probability of null3. This allows models
built with Infernal v1.0rc1->1.0.2 that use the old prior to have a
null3_omega of 1/32, and for those built with the new prior to be
1/65536. <null3_omega> is now stored in the CM file. For models built
with 1.0->1.0.2, it is absent, in which case it is set as 1/32.
Analagous changes have been made for a <null2_omega> value, the prior
probability of the null2 model, but null2 is current not used by
default.

Though the improvements on the benchmark seem to warrant the new
prior, all of these changes were structured to be easily reverted if
future experiments/observations suggests the old prior should not be
changed.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3366 db3e70e4-12e1-0310-b23f-81a6e3d09ba7