Still working on new cmsearch pipeline.
This commit modifies the CM data structure and CM file format to
include additional HMMs built from the input alignment to dcmbuild. A
single additional model is built using HMMER default entropy weighting
and priors with the --p7-add option to dcmbuild. This HMM is then read
and used in addition to the default maximum-likelihood (ML) HMM as
complementary filters in dcmsearch. Most changes to the code are
related to these additional p7 filters. I may eventually revert this
change (probably manually, i.e. not with an 'svn revert') after more
testing, but it was useful to commit it in the current state.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3455 db3e70e4-12e1-0310-b23f-81a6e3d09ba7