cm.c:           added 2 free calls to FreeCM() cm_io.c:        HMMER3/d format now fully supported (again, we need a better 	        way of reading HMM files eventually). cm_pipeline.c:  cosmetic changes to comment lines, only
dcmbuild.c:     
Added --p7-c* options for building HMMs from clusters of seed
sequences. These are analogous to the --c* options, but cause only an
HMM that will be used to filter in dcmsearch to be built from each
cluster, not a CM from each cluster. 

Also, --exp-real and --exp-fitlam are now default, and replaced with
the --exp-iid and --exp-fixlam options which turn them off. So, now by
default, realistic genome-like sequences are used for HMM calibration
(iid seqs with --exp-iid) and HMM E-value lambdas are fit by default
(approximately 0.693 is used if --exp-fixlam).


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3469 db3e70e4-12e1-0310-b23f-81a6e3d09ba7