Still working on new cmsearch pipeline.
Added several experimental options, most or all of which may be
removed soon. --glen, --glN, --glX, and --glstep lower the p7 glocal
forward filter P-value threshold (gF3) for long (consensus length >
200) models. By default, gF3 is halved for each 100 nt a models clen
exceeds 200, until 500. This gives about a 15% speedup on the rmark3
benchmark with no appreciable loss in sensitivity, though it must
decrease sensitivity in some 'real-life' cases.  The motivation is to
save time on larger models which take disproportiantely long in the
CM stages due to the high complexity of CYK/Inside.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3501 db3e70e4-12e1-0310-b23f-81a6e3d09ba7