Still working on new cmsearch pipeline.
Cleaned up/simplified code related to model building. From dcmbuild
(which will eventually become cmbuild): removed many of the options
related to building additional p7 HMMs as filters. Including all
related to clusters on the input sequences as these did not improve
performance on rmark3, while significantly increasing running times of
searches (see
~nawrockie/notebook/11_0224_inf_dcmsearch_p7s_from_clusters/00LOG for
results).

The new default strategy for model building is to build a single
additional HMM and include it in the CM file. It is built using HMMER
default transition priors for DNA/RNA HMMs and marginalized emissions
from a CM built to achieve a HMM mean match state relative entropy of
0.38 (paired positions are marginalized). Using this model in
combination with the ML p7 HMM built entirely from the CM (that is,
using two HMMs as filters) gives the best performance on the benchmark
of any combination of 2 mono-RNA HMM filters I've tried. di-RNA filters
still do very slightly better (about 2-3 MER), but the difference is
so slight that I've decided that, given their added complexity, di-RNA
filters are not worth committing to the trunk (they were implemented
in ~nawrockie/notebook/10_0901_inf_dinuc but never committed). 

Using only this model as a single filter gives the best single filter
results I've seen, and are performance is so close to the double
filters (4 MER or so) that I think I'll probably make mono-filtering
default.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3503 db3e70e4-12e1-0310-b23f-81a6e3d09ba7