esl_msa_Set*() functions now take an extra argument <n> for the string or buffer length. Optional (-1) if string; mandatory if unterminated buffer.
I couldn't test this change; I can't get Infernal to
build. Specifically, 'make clean' fails. There are (at least) two
copies of Easel in here. I think the autoconf script is only building
Makefiles for one of them, but the HMMER Makefile is trying to recurse
into the other. I think we need to revise so that HMMER configuration
is flexible about where Easel is (i.e. it can be ../easel, not easel/,
relative to H3).



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3538 db3e70e4-12e1-0310-b23f-81a6e3d09ba7