Overhaul of Infernal's build strategy to be consistent with the current build strategy in HMMER. This was done by literally copying the analogous Makefile.in and configure.ac files from HMMER and minimally changing them to work with Infernal (e.g. infernal/src/Makefile.in is based on hmmer/src/Makefile.in).
Thus, the Infernal build strategy is now consistent with HMMER as of
HMMER's 3540 revision.

This brings in nice-looking 'make' output stolen from Git, as well
as an overhaul of how the impl_sse code is handled (see HMMER revision
3513 for details).

As part of the overhaul, the test programs in testsuite/ were moved to
src/ to be more consistent with HMMER organization. These are now
compiled by a 'make' in infernal/src/'. The sqc files that use these
tests have been updated accordingly.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3543 db3e70e4-12e1-0310-b23f-81a6e3d09ba7