MPI dcmsearch now works, at least anecdotally. Statistics (number of residues surviving each filter) are slightly different from serial and threaded mode, which needs to be understood and (probably) fixed.
MPI dcmsearch requires an SSI index file for the database to run. This
makes it possible for workers to receive a work unit that tells it
which sequence(s) and residue boundaries to search, and quickly fetch
that (sub)sequence and do the search. Without SSI, it would need to
position the sqfile at the beginning of the first desired sequence,
requiring it to chew through potentially a lot of sequence until it
gets to the desired subsequence. I could try and position it at
subsequence starts without SSI, but that's not really supported by
Easel as esl_sqio_ReadWindow() expects the caller to sequentially read
complete sequences. Or I could actually send the sequence to the
workers (which old cmsearch did!), but we don't want to be sending all
that data around. SSI is a good solution, with the only drawback being
that users now need to index their files a la BLAST.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3560 db3e70e4-12e1-0310-b23f-81a6e3d09ba7