Added cmpress. It 'works', as in it compresses CM databases, but cmscan is broken now, because it hasn't yet been updated to work on pressed databases.
Changed how the HMMs filters that are embedded within CM files are
read. Previously, a P7_HMMFILE object was being created, opened and
closed for each CM read. This was wasteful, and slow. Now, a
P7_HMMFILE is part of the CM_FILE data structure, it is created/opened
once when CM_FILE is created/opened and used for all P7 filters in the
CM file. New way is cleaner and faster. Reading all 1973 Rfam 10.1 CMs
and p7 filters from a binary CM file dropped from ~0.5s to ~0.2s.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3603 db3e70e4-12e1-0310-b23f-81a6e3d09ba7