cmpress: cmpress now creates five binary files:   1. <cmfile>.i1m: complete CMs (and p7 filter HMMs, currently)   2. <cmfile>.i1i: SSI index for <cmfile>.i1m file.   3. <cmfile>.i1h: complete p7 filter HMMs.   4. <cmfile>.i1f: MSV filter part of optimized p7 profiles.   5. <cmfile>.i1p: remainder of optimized p7 profiles.
As of now, there can be more than one p7 filter for each CM, but I am
strongly considering enforcing a single p7 filter for each CM as this
would significantly simplify the code and multiple filters are only
used if certain command-line options are used. 

The .i1h file is identical to a HMMER3 binary file (.h3m) and Infernal
can use the HMMER input/output functions to deal with it. Similarly,
the .i1p file is identical to a .h3p file. The .i1f file is nearly
identical to the .h3f file, but includes a few pieces of
Infernal-specific data prepended to the HMMER3 format.

cmscan: 
Updated to only work on cmpress'ed CM database files. Still in
construction, currently only serial mode works and threaded code is
not even compiled. 

Given a cmpress-generated set of files, cmscan will read each 'model'
in several steps, but only as necessary, stopping and moving to the
next model when zero hits in the query sequence have survived past a
given filter stage. The model reading steps are, for a given CM:

1. MSV part of >=1 relevant optimized p7 filters from .i1f (mandatory)
2. Remainder of optimized p7 filters from .i1p (if a hit passes F1b)
3. Full p7 filters from .i1h (for generic profile, if hit passes F3b)
4. Full CM from .i1m (if a hit passes F5b)



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3607 db3e70e4-12e1-0310-b23f-81a6e3d09ba7