Major change in search/scan pipeline:
Each CM must have a single p7 HMM filter, stored as cm->fp7. This HMM
is necessarily saved in the CM file, in HMMER format. The filter HMM
can be the ML p7 HMM (the maximum likelihood HMM built to be as
similar as possible to the CM) but this must be dictated at the
cmbuild step. By default, it is a p7 HMM built using Infernal emission
priors, HMMER transition priors and a target mean match state entropy
of 0.38 (this strategy that optimizes performance on rmark3).

This change includes a minor modification to the CM file format, which
no longer has 'NAP7' lines and doesn't include E-value statistics for
ML p7 HMMs. 

Previously, the CM file could contain 0 or 1 'additional p7 HMMs' as
well as E-value statistics for the ML p7 HMM. cmsearch and cmscan
could use the ML p7 HMM, the additional P7 HMM, or both to filter. By
default only the additional HMM was used because that strategy gave
the best rmark3 results. 

This change simplifies CM_PIPELINE functions, which now necessarily
use a single p7 HMM, instead of needing to be flexible to use either
one or two HMMs. cmsearch and cmscan code is now cleaner as
well. Also, cmbuild is now faster (about 2X) because only one HMM must
be calibrated.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3610 db3e70e4-12e1-0310-b23f-81a6e3d09ba7