Top-level and src/-level 'make dev' now builds benchmarks and unit tests, and 'make check' builds unit tests for Infernal. Only one unit test and benchmark currently exist, for cm_tophits.
That benchmark helped uncovered a bug in how overlapping (duplicate)
hits were detected and removed, which is now fixed (hits on reverse
complement were being improperly sorted for the implemented duplicate
removal). The tophits benchmark also revealed that with the present
strategy it is not faster to sort tophits lists prior to merging them
in most cases, so a simpler method of merging those lists that simply
appends copies of all hits from one unsorted list to the other
unsorted list has been implemented.

Also, a <in_rc> parameter has been added to the CM_HIT structure, with
value TRUE if the hit is on the reverse strand, FALSE if
not. Previously, the code relied on <start> and <stop> to define
strand (forward strand if start < stop, else reverse strand), but that
doesn't work for hits that span a single residue. Such hits were
always defined as being on the reverse strand previously, and although
such hits are vanishingly rare if not impossible on normal search
conditions, the improper handling of them bothered me.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3611 db3e70e4-12e1-0310-b23f-81a6e3d09ba7