Simplified CM_FILE data structure by removing the P7_HMMFILE <gfp> pointer. All P7_HMMFILE manipulation is now handled by the other P7_HMMFILE in CM_FILE, <hfp>.
Previously, two P7_HMMFILEs were necessary to handle the two cases of
a CM file being pressed and not being pressed. If pressed, a separate
binary file of HMM filters existed (with suffix .i1h) because not all
p7 filters were necessarily written to the CM file and therefore had
to be written to a separate file since we needed the ability to read
them independently of the CM. However, with the recent change
requiring that all CMs have a single p7 filter that must be written to
the CM file, the .i1h file and <gfp> for handling it are no longer
necessary (cmpress no longer creates a .i1h file).

Most of the files listed as modified are minimally changed to be
consistent with a new cm_file_Read() function prototype. That function
now includes a <read_fp7> flag, specifying whether the p7 filter
should be read for the CM. In most cases it is TRUE, but in a scan
pipeline originating from cmscan, it is FALSE. In this case we've
previously read the p7 filter for the CM and a hit in the query
sequence has survived that filter, so we don't have to read the p7
filter again.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3616 db3e70e4-12e1-0310-b23f-81a6e3d09ba7