Updated esl_msa* and esl_msafile* calls to be consistent with the new API Sean introduced in revision 715 of Easel.
Update was straightforward, except I had to modify
cm_parsetree.c::Parsetrees2Alignment() to not use msa->abc (for
identifying gaps) when msa is in text mode (this is a similar problem
to the one Sean detected and fixed in the esl-ssdraw miniapp with
revision 715). It's clear to me now that using msa->abc in a text mode
is a no-no, because whether it is NULL or not is used as a flag by
esl_msa* functions to determine if the msa is digitized or not. (Didn't
that used to be an explicit flag: msa->flags & eslMSA_DIGITAL?)


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3633 db3e70e4-12e1-0310-b23f-81a6e3d09ba7