Modified how hits are aligned in the new pipeline. Previously, any hit i..j in a envelope 1..i..j..L that survived the final search stage of HMM banded scanning Inside was aligned with HMM banded optimal accuracy. The HMM bands from the final stage Inside scanner of 1..L were not used, but rather bands were recalculated for alignment of precisely i..j. In very rare cases, all of which I've seen occur with truncated sequences, the resulting Inside parse of i..j had a bit score much lower than the final stage bit score using HMM banded scanning Inside (by default). This occurs if the i..j alignment bands exclude the hit found by the scanner, which had different bands. In this case, the reported alignment would often look like crap even though the hit bit score (derived by the scanning Inside) was significant.
The current solution is to use the Inside scanning bands for the hit
alignment, thus guaranteeing that a hit with the same Inside score is
found during alignment, and its optimally accurate parse is reported.
Unfortunately, the scanning bands are often looser and this can cause
a overall slowdown of 10-30% for cmsearch on rmark3, depending on the
strictness of the filters. Also, the matrices used for alignment
(which requires two HMM banded DP matrices plus a shadow matrix) are
much larger in some cases now (2-3X bigger in extreme cases).

I'm not completely satisfied with the current solution, and may come
back and reevaluate this in the future. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3646 db3e70e4-12e1-0310-b23f-81a6e3d09ba7