cm_dpsearch_truncyk.c: Found and fixed a bug involving using QDBs on B states. ref: ~nawrockie/notebook/11_0816_inf_banded_trcyk/00LOG.
cmemit.c: Cleaned up another case of using msa->abc that should've
been fixed in previous commit.

trcyk.c: fixed a typo; an extra ';' that I had added when I updated
this file for use with the new CM file format. This was causing the
program to report an error with reading the CM file even if there was
no problem.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3657 db3e70e4-12e1-0310-b23f-81a6e3d09ba7