Adding cm_dpsearch_trunc.c::FastTrCYKScan() based on FastCYKScan(). Unfortunately, trCYK is different enough from CYK that the speed improvements seen between FastCYKScan() and RefCYKScan() (about 30%) are much lower in the Tr (truncated) version (less than 5%).
I plan to remove FastTrCYKScan() in the next commit, but I wanted to
commit it once so its code was in the subversion repository in case
anyone wants to revisit it to optimize it. I decided to scrap it
because it's marginal speed improvement doesn't outweigh its
disadvantages: it is long (three times as many lines as
RefTrCYKScan()) and difficult to debug/maintain. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3663 db3e70e4-12e1-0310-b23f-81a6e3d09ba7