Fixed and logged bug i27 in BUGTRAX. Marginal emission scores used by all truncated CYK implementations were being incorrectly calculated due the use of the esl_vec_FLogsum() function on log odds scores in base 2 (that function assumes log base e). This has been fixed by rewriting how marginal scores are calculated and moving their calculation to CMLogoddisfy(). Now whenever the CM bit scores are valid, the marginal bit scores are as well (as opposed to requiring a special call to SetMarginals()).
The alphabet.c::LeftMarginalScore() and
alphabet.c::RightMarginalScore() functions contained the incorrect
usage of the esl_vec_FLogsum() function. These were renamed 
by adding '_reproduce_bug_i27' to their names and moved to
truncyk.c. I did this to enable reproduction of the benchmark from
Kolbe, Eddy 2009, but I'm not sure this is important since the version
that corresponds to that publication already enables reproduction.
I may remove this in a future commit.

The benchmark-trunc-search test program optionally compiled with
cm_dpsearch_trunc.c includes the option --i27 to use the old method
for calculating the marginal scores. 

Most of the files that were modified were to remove calls to
LeftMarginalScore() and RightMarginalScore(). They can simply use the
pre-calculated scores in cm->lmesc and cm->rmesc vectors now.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3666 db3e70e4-12e1-0310-b23f-81a6e3d09ba7