Most files were modified to use the function prototype for DispatchAlignments(), cp9_HMM2ijBands(), ij2d_bands() and cp9_ShiftCMBands() all of which now take a <do_trunc> argument, which is TRUE if the HMM bands will be used to contrain a trCYK/trInside/trOutside matrix, and FALSE otherwise.  Truncated alignment requires special considerations when constructing the bands, usually leading to wider bands for some states, which are unnecessary if the bands will be used to constrain a standard non-truncated CYK/Inside/Outside algorithm.
cm_dpsearch_trunc.c:

Updated FastTrCYKScanHB() so it finds optimal parse for all tested
examples on homologous or emitted sequences and the vast majority of
random sequences. More rigorous testing to come. Unfortunately (and
this was not immediately obvious) HMM bands need to be determined
differently to allow for marginal L and R hits. For example, if the
left consensus position modeled by a base pair is determined to be
unlikely to emit any residue in the target, we still need to allow
that base pair state to be reachable for right marginal hits. The
analogous case is true for the right position and left marginal hits,
and for right marginal hits in MATL-modeled positions and left
marginal hits in MATR-modeled positions. Currently, only bands for
MATP_MP, MATR_MR, and MATL_ML states are extended to allow for
marginal hits (i.e. bands for MATP_MR, MATP_ML, inserts, deletes are
not extended). This should be okay; although I believe its possible to
miss the optimal parse in rare cases, the difference in bit score and
parsetree should be essentially insignificant.

Due to the expansion of bands to allow for marginal hits, the number
of cells excluded from HMM bands calculated for a standard CYK parse
versus bands calculated for a trCYK parse is roughly 3X higher. This
means HMM banded trCYK could require about 9X more time and memory
than standard CYK (because 3 full DP matrices are used instead of
1). But in reality, HMM banded trCYK seems roughly only 3-5X slower
than standard, although 9X more memory is required. For large models
(RNase P, SSU), the time required for the banded CYK tends to be equal
or less than the time required to calculate the bands, meaning there's
an overall slow down of only ~2X for band calculation plus CYK.

The tightness of the bands and speed of CYK could be improved by
determining and enforcing different HMM bands for each of the J,L,R,T
matrices, because the wider bands for any state relative to standard
CYK are only necessary for the L or the R matrix (never J and never
both L and R), but that would require even more convoluted HMM banded
CYK/Inside/Outside/Posterior functions with about 3X as much overhead
in time and code for edge-case checking, and I'm not sure they're
worth the trouble to write/debug.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3674 db3e70e4-12e1-0310-b23f-81a6e3d09ba7